+++ week4/controllers/people.js
@@ -1,42 +1,65 @@
-const products = [
-  {
-    id: 1,
-    name: 'albany sofa',
-    image:
-      'https://dl.airtable.com/.attachments/6ac7f7b55d505057317534722e5a9f03/9183491e/product-3.jpg',
-    price: 39.95,
-    desc: `I'm baby direct trade farm-to-table hell of, YOLO readymade raw denim venmo whatever organic gluten-free kitsch schlitz irony af flexitarian.`,
-  },
-  {
-    id: 2,
-    name: 'entertainment center',
-    image:
-      'https://dl.airtable.com/.attachments/da5e17fd71f50578d525dd5f596e407e/d5e88ac8/product-2.jpg',
-    price: 29.98,
-    desc: `I'm baby direct trade farm-to-table hell of, YOLO readymade raw denim venmo whatever organic gluten-free kitsch schlitz irony af flexitarian.`,
-  },
-  {
-    id: 3,
-    name: 'albany sectional',
-    image:
-      'https://dl.airtable.com/.attachments/05ecddf7ac8d581ecc3f7922415e7907/a4242abc/product-1.jpeg',
-    price: 10.99,
-    desc: `I'm baby direct trade farm-to-table hell of, YOLO readymade raw denim venmo whatever organic gluten-free kitsch schlitz irony af flexitarian.`,
-  },
-  {
-    id: 4,
-    name: 'leather sofa',
-    image:
-      'https://dl.airtable.com/.attachments/3245c726ee77d73702ba8c3310639727/f000842b/product-5.jpg',
-    price: 9.99,
-    desc: `I'm baby direct trade farm-to-table hell of, YOLO readymade raw denim venmo whatever organic gluten-free kitsch schlitz irony af flexitarian.`,
-  },
-]
-const people = [
-  { id: 1, name: 'john' },
-  { id: 2, name: 'joy' },
-  { id: 3, name: 'dilan' },
-  { id: 4, name: 'anna' },
-  { id: 5, name: 'emma' },
-]
-module.exports = { products, people }
+
+let { people } = require('../data')
+
+const getPeople = (req, res) => {
+  res.status(200).json({ success: true, data: people })
+}
+
+const createPerson = (req, res) => {
+  const { name } = req.body
+  if (!name) {
+    return res.status(400).json({ success: false, msg: 'please provide your name' })
+  }
+  res.status(201).send({ success: true, data: [...people, name] })
+}
+
+const createPostmanPerson = (req, res) => {
+  const { name } = req.body
+  if (!name) {
+    return res
+      .status(400)
+      .json({ success: false, msg: 'please provide name value' })
+  }
+  res.status(201).json({ success: true, data: [...people, name] })
+}
+
+const updatePostmanPerson = (req, res) => {
+  const { id } = req.params
+  const { name } = req.body
+
+  const person = people.find((person) => person.id === Number(id))
+
+  if (!person) {
+    return res
+      .status(404)
+      .json({ success: false, msg: `no person with id: ${id}` })
+  }
+  const newPeople = people.map((person) => {
+    if (person.id === Number(id)) {
+      person.name = name
+    }
+    return person
+  })
+  res.status(200).json({ success: true, data: newPeople })
+}
+
+const deletePerson = (req, res) => {
+  const person = people.find((person) => person.id === Number(req.params.id))
+  if (!person) {
+    return res
+      .status(404)
+      .json({ success: false, msg: `no person with id ${req.params.id}` })
+  }
+  const newPeople = people.filter(
+    (person) => person.id !== Number(req.params.id)
+  )
+  return res.status(200).json({ success: true, data: newPeople })
+}
+
+module.exports = {
+  getPeople,
+  createPerson,
+  createPostmanPerson,
+  updatePostmanPerson,
+  deletePerson
+}
